Разделяй и властвуй (separating utils.js)
