+++ src/main.js
@@ -11,7 +11,7 @@ import Popup from './view/popup.js';
 
 import {generateFilm} from './mock/film.js';
 import {generateStat} from './mock/stat.js';
-import {render, RenderPosition} from './utils.js';
+import {render, RenderPosition} from './utils/render.js';
 
 const FILM_COUNT = 22;
 const FILM_COUNT_PER_STEP = 5;
@@ -46,19 +46,19 @@ const renderFilmCard = (container, filmCard) => {
   render(container, filmCardElement.getElement(), RenderPosition.BEFOREEND);
 };
 
-render(header, new UserRank(userRank).getElement(), RenderPosition.BEFOREEND);
-render(main, new Films().getElement(), RenderPosition.BEFOREEND);
+render(header, new UserRank(userRank), RenderPosition.BEFOREEND);
+render(main, new Films(), RenderPosition.BEFOREEND);
 
 const films = main.querySelector('.films');
 
 if (filmCards.length > 0) {
-  render(main, new Sort().getElement(), RenderPosition.AFTERBEGIN);
-  render(main, new Nav(stat).getElement(), RenderPosition.AFTERBEGIN);
-  render(films, new FilmsList().getElement(), RenderPosition.BEFOREEND);
+  render(main, new Sort(), RenderPosition.AFTERBEGIN);
+  render(main, new Nav(stat), RenderPosition.AFTERBEGIN);
+  render(films, new FilmsList(), RenderPosition.BEFOREEND);
 
   const filmsList = films.querySelector('.films-list');
 
-  render(filmsList, new FilmsListContainer().getElement(), RenderPosition.BEFOREEND);
+  render(filmsList, new FilmsListContainer(), RenderPosition.BEFOREEND);
 
   const filmsListContainer = filmsList.querySelector('.films-list__container');
 
@@ -69,7 +69,7 @@ if (filmCards.length > 0) {
   if (filmCards.length > FILM_COUNT_PER_STEP) {
     let renderedFilmCount = FILM_COUNT_PER_STEP;
 
-    render(filmsList, new ShowMoreButton().getElement(), RenderPosition.BEFOREEND);
+    render(filmsList, new ShowMoreButton(), RenderPosition.BEFOREEND);
 
     const showMoreButton = filmsList.querySelector('.films-list__show-more');
 
@@ -77,7 +77,7 @@ if (filmCards.length > 0) {
       evt.preventDefault();
       filmCards
         .slice(renderedFilmCount, renderedFilmCount + FILM_COUNT_PER_STEP)
-        .forEach((filmCard) => render(filmsListContainer, new FilmCard(filmCard).getElement(), RenderPosition.BEFOREEND));
+        .forEach((filmCard) => render(filmsListContainer, new FilmCard(filmCard), RenderPosition.BEFOREEND));
 
       renderedFilmCount += FILM_COUNT_PER_STEP;
 
@@ -93,9 +93,9 @@ if (filmCards.length > 0) {
     .slice(0, FILM_EXTRA_COUNT)
     .filter((film) => film.totalRating);
   if (topRatedfilmCards.length) {
-    render(films, new FilmsList('Top rated', 'top-rated').getElement(), RenderPosition.BEFOREEND);
+    render(films, new FilmsList('Top rated', 'top-rated'), RenderPosition.BEFOREEND);
     const topRatedfilmsList = films.querySelector('#top-rated');
-    render(topRatedfilmsList, new FilmsListContainer().getElement(), RenderPosition.BEFOREEND);
+    render(topRatedfilmsList, new FilmsListContainer(), RenderPosition.BEFOREEND);
     const topRatedfilmListContainer = topRatedfilmsList.querySelector('.films-list__container');
     topRatedfilmCards.forEach(
       (topRatedfilmCard) => renderFilmCard(topRatedfilmListContainer, topRatedfilmCard));
@@ -107,16 +107,16 @@ if (filmCards.length > 0) {
     .slice(0, FILM_EXTRA_COUNT)
     .filter((film) => film.comments.length);
   if (mostCommentedfilmCards.length) {
-    render(films, new FilmsList('Most commented', 'most-commented').getElement(), RenderPosition.BEFOREEND);
+    render(films, new FilmsList('Most commented', 'most-commented'), RenderPosition.BEFOREEND);
     const mostCommentedfilmsList = films.querySelector('#most-commented');
-    render(mostCommentedfilmsList, new FilmsListContainer().getElement(), RenderPosition.BEFOREEND);
+    render(mostCommentedfilmsList, new FilmsListContainer(), RenderPosition.BEFOREEND);
     const mostCommentedfilmListContainer = mostCommentedfilmsList.querySelector('.films-list__container');
     mostCommentedfilmCards.forEach(
       (mostCommentedfilmCard) => renderFilmCard(mostCommentedfilmListContainer, mostCommentedfilmCard));
   }
 } else {
-  render(main, new Nav(stat).getElement(), RenderPosition.AFTERBEGIN);
-  render(films, new FilmsList(EMPTY_LIST_TITLE).getElement(), RenderPosition.BEFOREEND);
+  render(main, new Nav(stat), RenderPosition.AFTERBEGIN);
+  render(films, new FilmsList(EMPTY_LIST_TITLE), RenderPosition.BEFOREEND);
 }
 
-render(footerStatistics, new FooterStatistics(filmCards.length).getElement(), RenderPosition.BEFOREEND);
+render(footerStatistics, new FooterStatistics(filmCards.length), RenderPosition.BEFOREEND);
